logos, text color, gradient color, added summaries
Basically put each teams logo in their box. Added our new logo on top.
text colors were hard to see as they were vibrant. Changed it to match the new logos that are better to see.
changed the color of the background for the text to be greyish blue. Also changed the article background color to a greyish blue.
added summaries into each team and intro.
found a bug that you can basically move the window more right, which will reveal a bunch of stuff that is not needed.
There is a greyish-blue bar below the team image on the main screen.